Change the theme and toggle dark mode in shadcn
This was done after visiting https://ui.shadcn.com/docs/theming and copy the code for the selected theme from there and replace the theme that was put as default theme.

Now we can use some of these variables in the first @layer base in other parts in our project.