Setup form component and types
This is a little bit complicated, but it's better idea do it in that way. We spares lines of code but (for example) that the moment we'll need more interaction, as in the moment when submit button is set to pending while the user is submitting the form, we'll find ourselves adding more code. That's why this is the better approach. We  installed form input for that goal with npx shadcn@latest add form input.

We extract the types in separate components so that we could reuse it. Use Zod library for this goal.