Finish the create job form
We use useMutation (the react query method) so that we could add the job.
Install the toast library which we forgot to do before the last push.
Now when we create a job, it creates the job and navigate us to the jobs page. The job is thereafter saved in prisma.