Setup all jobs function
Page and limit for now are hardcoded. We'll return to them later.

Explanation of the new code:  Prisma.JobWhereInput is a specific type that prisma uses for that where object. And for the time being, we just set up the clerkId. That's our initial setup.
But then if the search is not empty string or undefined, then we grab the existing values from whereClause and we set up our logic with the OR. So essentially if there's something in search, we look for all the jobs where the position contains this value or the company contains this value.
And when it comes to jobStatus, the main idea is the same. I check whether it's true and if it's not equal to all. So if it's equal to all, then i don't include it, because i just wanna get all of the jobs.
After this, in the whereClause object we destructure the current whereClause, whatever it is, and then we just set status equal to a jobStatus. And then instead of hard-coding over here, we have this whereClause in place. And we order based on createdAt and set it equal to descending.
If we're successful, we return jobs, which is array of JobType, and again, for now we're just hard-coding the values for count and page. And if we have an error - we return an empty array with jobs.