Setup the jobs list component
Since react query caches everything, we wanna add also those values over here (search ?? "", jobStatus, pageNumber) and that's why in the page in jobs we were providing these default ones in queryKey - because now in the JobsList, one by one, we'll set it equal to the actual constants.
THE REASON why we wanna add these constants over here is because we wanna refetch if the value is changing, otherwise it's going to be cached.
The reason why we pass them in the getAllJobsAction - we wanna use them when we're querying the database with Prisma. And that's why in the page we setup those default values in the query ("jobs", "", "all", 1). This always gonna be the case if i don't have any search params.
Now when we hit enter with empty search field and status "all", the URL is
http://localhost:3000/jobs?search=&jobStatus=all